dodata strana za sastojke, dodat edit za sastojke, dodati novi kokteli i sredjen view za koktele

+++ app/Models/Sastojak.php
@@ -30,7 +30,7 @@ class Sastojak extends Model
     {
         return [
             'id' => 'integer',
-            'kolicina' => 'decimal',
+            'kolicina' => 'float',
         ];
     }
